Remove reminder label in finish method
Add module global string var REMIND_TO_FINISH with text to show when
the user does not use the finish method.

When the clear method is called, and thus also at init time, a label
with REMIND_TO_FINISH text is added to the grid. This is only removed
again when the user calls the finish method.